refactor(maintainer): 1.fix some test bug 2.TSSManager voteTxIn and  voteTxOut fanction change to batch submit 3.Maintainers -> maintainer address and validator independent of each other

+++ maintainer/contracts/libs/Types.sol
@@ -31,14 +31,14 @@ struct BridgeItem {
 struct TxInItem {
     bytes32 orderId;
     BridgeItem bridgeItem;
-    uint128 height;
+    uint64 height;
     bytes refundAddr;
 }
 
 struct TxOutItem {
     bytes32 orderId;
     BridgeItem bridgeItem;
-    uint128 height;
+    uint64 height;
     uint128 gasUsed;
     address sender;
 }
